Battle logic and code isn't 100 percent perfect but should be a good start for it.
