+++ IslandOdyssey/MonsterFile.txt
@@ -1,7 +1,7 @@
-Shifty the Coconut Crab~ Massive coconut crab. His shadiness is matched almost by his skill of thievery. He is the king of decapods. He is very possessive of his “belongings” and will instantly panic once anything goes missing.~10~20~50~Beach Ball~Puzzle Key
-Dasami the Komodo Dragon~Massive dragon that doesn’t breathe fire (luckily). Looks menacing, but something about them gives off a very friendly aura. Dasami does use venom whenever needed though, so be careful!~25~100~60~Maracas~None
-Angry Alligator~A baby gator that believes they are the size of a full-adult gator. Will attack anything that moves – luckily, they are angrier than they are powerful.~50~5~80~Flame Thrower~Alligator Meat, Puzzle Key
+Shifty the Coconut Crab~ Massive coconut crab. His shadiness is matched almost by his skill of thievery. He is the king of decapods. He is very possessive of his “belongings” and will instantly panic once anything goes missing.~10~20~50~Beach Ball
+Dasami the Komodo Dragon~Massive dragon that doesn’t breathe fire (luckily). Looks menacing, but something about them gives off a very friendly aura. Dasami does use venom whenever needed though, so be careful!~25~100~60~Maracas
+Angry Alligator~A baby gator that believes they are the size of a full-adult gator. Will attack anything that moves – luckily, they are angrier than they are powerful.~50~5~80~Flame Thrower
 Spiders~Spiders that cannot be killed by crushing them – there's simply too many. Only something that can defeat them all in one fell swoop is effective.~10~100~30~Flame Thrower~Puzzle Key
-Roger the Ravenous Sea Lion~A sea lion that hasn’t seen a good meal in a few weeks – until you showed up. He is so hungry he will clumsily bite anything he sees – perhaps to his doom.~25~30~20~Giant Fish Net~Knife, Coconut
-Sly the Anaconda~A slippery and slick snake, Sly the anaconda is a master of disguise - she just loves to give hugs... until you die. You’ve got to learn how to cut her off.~30~25~30~Knife~Puzzle Key
-Zamrah the Minotaur~A mythical creature known only to those braves enough to call themselves adventurers. Immune to effects from items. Legend has it that the brave adventurer Harmaz once defeated a Minotaur in one fell swoop with a mighty machete strike to the head.~50~50~70~Machete~None
+Roger the Ravenous Sea Lion~A sea lion that hasn’t seen a good meal in a few weeks – until you showed up. He is so hungry he will clumsily bite anything he sees – perhaps to his doom.~25~30~20~Giant Fish Net
+Sly the Anaconda~A slippery and slick snake, Sly the anaconda is a master of disguise - she just loves to give hugs... until you die. You’ve got to learn how to cut her off.~30~25~30~Knife
+Zamrah the Minotaur~A mythical creature known only to those braves enough to call themselves adventurers. Immune to effects from items. Legend has it that the brave adventurer Harmaz once defeated a Minotaur in one fell swoop with a mighty machete strike to the head.~50~50~70~Machete
