+++ IslandOdyssey/src/main/java/Monster.java
@@ -1,6 +1,7 @@
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.Scanner;
+import java.util.SortedMap;
 
 public class Monster {
     //Created by Joseph
@@ -13,15 +14,17 @@ public class Monster {
     Item item;
     boolean isDefeated = false;
 
+    public Monster(){
+
+    }
     //From Joseph
-    public Monster(String name, String monsterDescription, int hitPoints, int strength, int attackChance, String weakness, Item item){
+    public Monster(String name, String monsterDescription, int hitPoints, int strength, int attackChance, String weakness){
         this.name = name;
         this.monsterDescription = monsterDescription;
         this.hitPoints = hitPoints;
         this.strength = strength;
         this.attackChance = attackChance;
         this.weakness = weakness;
-        this.item = item;
     }
 
     //Joseph
@@ -30,9 +33,10 @@ public class Monster {
         return item;
     }
     //Joseph
-//    public void addItemToMonster(){
-//
-//    }
+   public void addItemToMonster(Item monsterItem){
+        Monster monster = new Monster();
+        monster.addItemToMonster(monsterItem);
+   }
 
     //Joseph
     public void takeHit(int hit){
@@ -49,9 +53,41 @@ public class Monster {
 
     }
     //Joseph
-//    public void battle(){
-//
-//    }
+    public void battle(){
+        System.out.println("This is " + getName());
+        Monster battleMonster = new Monster();
+        Player player = new Player("RoomFile.txt","ItemFile.txt","MonsterFile.txt","PuzzleFile.txt");
+        System.out.println("Do you want to attack the monster or ignore the call for battle!");
+        Scanner input = new Scanner(System.in);
+        String action = input.nextLine();
+        while(!action.equalsIgnoreCase("Ignore")){
+            if(action.equalsIgnoreCase("examine")){
+                battleMonster.getMonsterDescription();
+            }
+            if(action.equalsIgnoreCase("Attack") || action.equalsIgnoreCase("A")){
+                battleMonster.takeHit(player.getAttackPower());
+                player.takeHit(battleMonster.getStrength());
+
+                if(action.equalsIgnoreCase("Use")){
+                    player.use(item);
+                    if (item.equals(weakness)){
+                        String endItem = weakness;
+                        battleMonster.isWeakTo(endItem);
+                        break;
+                    }
+                    else if(item.equals("Maracas")){
+                        player.getPreviousRoom();
+                        break;
+                    }
+                }
+                battleMonster.whenDefeated();
+                break;
+
+            }
+        }
+        player.getPreviousRoom();
+
+    }
     //Joseph
     public boolean isWeakTo(String itemName){
         if (itemName.equals(weakness)){
@@ -61,7 +97,7 @@ public class Monster {
 
         return false;
     }
-    //
+    //Joseph
     public void whenDefeated(){
         isDefeated = true;
     }
