+++ README.txt
@@ -1,0 +1,21 @@
+MonsterFile.txt
+This text file is the basis of important variables that is parse into the map for the monster class.
+In total there are six variables that the monster class will contain, these are name, monsterDescription, hitPoints, strength, attackChance, and weakness.
+
+<name>
+A string variable which will provide the name of a monster.
+
+<monsterDescription>
+It's a string which parse a description of a monster when player examines interaction with the monster
+
+<hitPoints>
+This int variable in the amount of health the monster.
+
+<strength>
+It's an int variable about the amount of damage the monster can put to the player when in combat.
+
+<attackChance>
+This is an int variable where the monsters have a percentage of success in hitting the player. When it's the monsters turn.
+
+<weakness>
+The string is meant to call an item where a specific item will have a weakness to. When enacted the monster's health will deplete to zero.
